Update CSS, add currency filter feature, modify cart slice
A significant overhaul of the CSS in the project was implemented, enhancing the overall UI/UX. A CurrencyFilter feature was added, which allows users to choose a currency. Modifications were made to the cartSlice.js file. Various pieces of information were added to the return action types of addItem and changeItemQuantity. Project setup files were adjusted and Redux package was integrated as well.